Reconfigured project so it uses packages insted of modules. Updated installation instructions in readme. Refactored .gitignore so it has the reccomended configuration for Intellij Spring Boot developement.
