+++ .idea/encodings.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/application/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/assessment/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/core/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/digital-library/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/digital-resources/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/interoperability/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/portal/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/reporting-statistics-visualization/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/src/main/resources" charset="UTF-8" />
+  </component>
+</project>
