+++ README.md
@@ -1,6 +1,6 @@
 # TeslaRIS-backend
 
 ## Installation
-Installation is very easy, first, position yourself into the directory that contains ```ApplicationStarter.java``` file, you can find it here: ```src/main/java/rs/teslaris/application/ApplicationStarter.java```. After that, simply run: ```mvn spring-boot:run``` which will run your application.
+Installation is very easy, first, position yourself into the project's root directory. After that, simply run: ```mvn spring-boot:run``` which will run your application.
 
-Alternatively, if you are using Intellij IDE, it is as easy as simply running ```ApplicationStarter.java``` file. 
+Alternatively, if you are using Intellij IDE, it is as easy as simply running ```TeslarisStarter.java``` file.
