--- application/src/main/java/rs/teslaris/application/ApplicationStarter.java
@@ -1,15 +1,0 @@
-package rs.teslaris.application;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackages = {"rs.teslaris"})
-@EnableAutoConfiguration
-public class ApplicationStarter {
-
-  public static void main(String[] args) {
-    SpringApplication.run(ApplicationStarter.class, args);
-  }
-
-}
