--- application/src/test/java/rs/teslaris/application/ApplicationTests.java
@@ -1,13 +1,0 @@
-package rs.teslaris.application;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
