--- core/src/test/java/rs/teslaris/core/CoreTests.java
@@ -1,9 +1,0 @@
-package rs.teslaris.core;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ContextConfiguration;
-
-@SpringBootTest
-public class CoreTests {
-}
