+++ pom.xml
@@ -1,22 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+
 	<modelVersion>4.0.0</modelVersion>
+
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
 		<version>2.7.9</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
+
 	<groupId>rs.teslaris</groupId>
-	<artifactId>parent</artifactId>
+	<artifactId>teslaris</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<packaging>pom</packaging>
+
+	<packaging>jar</packaging>
 	<name>parent</name>
 	<description>Parent project for TeslaRIS platform.</description>
+
 	<properties>
 		<java.version>11</java.version>
 	</properties>
+
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
@@ -74,17 +80,6 @@
 <!--			<scope>test</scope>-->
 <!--		</dependency>-->
 	</dependencies>
-
-	<modules>
-		<module>core</module>
-		<module>application</module>
-		<module>assessment</module>
-		<module>reporting-statistics-visualization</module>
-		<module>interoperability</module>
-		<module>portal</module>
-		<module>digital-resources</module>
-		<module>digital-library</module>
-	</modules>
 
 	<reporting>
 		<plugins>
