+++ src/main/java/rs/teslaris/teslaris/TeslarisStarter.java
@@ -1,0 +1,15 @@
+package rs.teslaris.teslaris;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@EnableAutoConfiguration
+public class TeslarisStarter {
+
+  public static void main(String[] args) {
+    SpringApplication.run(TeslarisStarter.class, args);
+  }
+
+}
