+++ src/main/java/rs/teslaris/teslaris/core/controller/DummyController.java
@@ -1,0 +1,15 @@
+package rs.teslaris.teslaris.core.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/dummy")
+public class DummyController {
+
+  @GetMapping
+  public String returnDummyData() {
+    return "TEST";
+  }
+}
