+++ src/main/resources/application.properties
@@ -1,0 +1,16 @@
+# SPRING
+spring.application.name=teslaris
+server.port=8081
+server.address=localhost
+
+# DATA SOURCE
+spring.datasource.url=jdbc:postgresql://localhost:5432/TeslaRISDB
+spring.datasource.username=${DB_USERNAME:postgres}
+spring.datasource.password=${DB_PASSWORD:root}
+
+# HIBERNATE
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.open-in-view=false
