+++ src/test/java/rs/teslaris/teslaris/TeslarisTests.java
@@ -1,0 +1,13 @@
+package rs.teslaris.teslaris;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TeslarisTests {
+
+  @Test
+  void contextLoads() {
+  }
+
+}
